refactor: 쿠폰 포맷팅 로직 유틸로 분리
