+++ src/features/coupon/ui/CouponCard.tsx
@@ -4,7 +4,7 @@ import * as S from './CouponCard.styles';
 import { useCouponsContext } from '../context/useCouponsContext';
 import { useSelectedCartItemsContext } from '../../cart/context/useSelectedCartItemsContext';
 import { useOrderContext } from '../../order/context/useOrderContext';
-import { isCouponApplicable } from '../utils';
+import { isCouponApplicable, formatCouponExpirationDate, formatCouponTimeRange, formatMinimumAmount } from '../utils';
 
 interface CouponCardProps {
   coupon: Coupon;
@@ -30,27 +30,6 @@ export default function CouponCard({ coupon }: CouponCardProps) {
     }
   };
 
-  const expirationDate = new Date(coupon.expirationDate ?? '');
-  const formattedExpirationDate = expirationDate.toLocaleDateString('ko-KR', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-
-  const formatTime = (timeString: string) => {
-    const [hours] = timeString.split(':').map(Number);
-    const period = hours < 12 ? '오전' : '오후';
-    const displayHours = hours === 0 ? 12 : hours > 12 ? hours - 12 : hours;
-    return `${period} ${displayHours}시`;
-  };
-
-  const formatTimeRange = () => {
-    if (!coupon.availableTime) return null;
-    const start = formatTime(coupon.availableTime.start);
-    const end = formatTime(coupon.availableTime.end);
-    return `${start}부터 ${end}까지`;
-  };
-
   return (
     <S.CouponCardContainer>
       <S.CouponCheckBox>
@@ -58,11 +37,11 @@ export default function CouponCard({ coupon }: CouponCardProps) {
         <S.CouponDescription>{coupon.description}</S.CouponDescription>
       </S.CouponCheckBox>
       <S.CouponValidInfoContainer>
-        <S.CouponValidInfo>만료일: {formattedExpirationDate}</S.CouponValidInfo>
+        <S.CouponValidInfo>만료일: {formatCouponExpirationDate(coupon.expirationDate)}</S.CouponValidInfo>
         {coupon.minimumAmount && (
-          <S.CouponValidInfo>최소 주문 금액: {coupon.minimumAmount?.toLocaleString()}원</S.CouponValidInfo>
+          <S.CouponValidInfo>최소 주문 금액: {formatMinimumAmount(coupon.minimumAmount)}</S.CouponValidInfo>
         )}
-        {coupon.availableTime && <S.CouponValidInfo>사용 가능 시간: {formatTimeRange()}</S.CouponValidInfo>}
+        {coupon.availableTime && <S.CouponValidInfo>사용 가능 시간: {formatCouponTimeRange(coupon)}</S.CouponValidInfo>}
       </S.CouponValidInfoContainer>
     </S.CouponCardContainer>
   );
